backend/requirements.txt
# Django and core dependencies
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==2.1.0

# PostgreSQL
psycopg2-binary==2.9.9

# Authentication and security
djangorestframework-simplejwt==5.3.1
django-ratelimit==4.1.0
bcrypt==4.1.1

# OpenAI integration
openai==1.6.1

# Environment and configuration
python-dotenv==1.0.0

# ASGI/WSGI servers
gunicorn==21.2.0
uvicorn[standard]==0.25.0
daphne==4.0.0

# WebSocket support for real-time chat
channels==4.0.0
channels-redis==4.1.0

# Redis for caching and WebSocket
redis==5.0.1

# Task queue (optional, for async AI processing)
celery==5.3.4

# Utilities
Pillow==10.1.0
uuid==1.30
python-dateutil==2.8.2

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.19.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==21.0.0

# Code quality
black==23.12.1
flake8==7.0.0
isort==5.13.2
pylint==3.0.3
mypy==1.8.0
django-stubs==4.2.7

# API documentation
drf-spectacular==0.27.0
drf-yasg==1.21.7

# Monitoring and logging
django-prometheus==2.3.1
sentry-sdk==1.39.1

# Data validation
marshmallow==3.20.1
pydantic==2.5.3

# Timezone support
pytz==2023.3.post1
PyJWT==2.8.0

